moved actuve members to guild resolver
